.github/scripts/duplicate_detection/requirements.txt
pygithub
ollama